Corrected typecast.

+++ coders/tiff.c
@@ -3718,7 +3718,7 @@ static MagickBooleanType WriteTIFFImage(const ImageInfo *image_info,
       predictor=PREDICTOR_FLOATINGPOINT;
     option=GetImageOption(image_info,"tiff:predictor");
     if (option != (const char * ) NULL)
-      predictor=(size_t) strtol(option,(char **) NULL,10);
+      predictor=(uint16) strtol(option,(char **) NULL,10);
     if (predictor != 0)
       (void) TIFFSetField(tiff,TIFFTAG_PREDICTOR,predictor);
     if ((image->x_resolution != 0.0) && (image->y_resolution != 0.0))
